test(evals): fold output_str JSON-contract checks into test_eval_dual_write (TH-4904)
Add 3 composition tests asserting output_str parses as valid JSON for
structured dict/list results when the output type is resolved off the
eval-template config. Drop the separate file in favour of the existing suite.
